Refactor(importCSVfile) #1
- View, ViewModel 분리
- loadCSV 함수를 DataFrame extension으로 정리
- Grid 컴포넌트의 최소버전이 16이상이므로 LazyVGrid로 변경
- Color Asset이 추가됨에 따라 background 컬러 적용을 RGB 값이 아닌 Assets 이름으로 변경
