+++ GGomVoca/GGomVoca/View/WordListView/JPWordListView/JPAddNewWordView.swift
@@ -108,7 +108,4 @@ struct JPAddNewWordView: View {
             }
         }
     }
-    
-    
-    
 }
